requirements.txt
alembic==1.9.4
aniso8601==9.0.1
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.0
Jinja2==3.1.2
jsonfy==0.4
jsonify==0.5
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
Pillow==9.4.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
